requirements.txt
#version for keycloack and python 3.10
appdirs==1.4.4
attrs==24.2.0
azure-common==1.1.28
azure-storage-blob==12.20.0
azure-storage-common==2.1.0
beautifulsoup4==4.12.3
black==24.4.2
certifi==2024.7.4
cffi==1.16.0
cfgv==3.4.0
chardet==5.2.0
clamd==1.0.2
click==8.1.7
cryptography==42.0.8
dateparser==1.2.0
defusedxml==0.7.1
distlib==0.3.8
Django==4.2.16
django-environ==0.11.2
django-filter==24.2
django-helusers==0.13.3
django-simple-history==3.7.0
django-storages==1.14.4
djangorestframework==3.15.2
djangorestframework-api-key==3.0.0
django-request-logging
drf-oidc-auth==3.0.0
ecdsa==0.19.0
filelock==3.15.4
flake8==7.1.0
future==1.0.0
helsinki-profile-gdpr-apI==0.2.0
huey==2.5.1
identify==2.6.0
idna==3.7
importlib-metadata==8.0.0
isodate==0.6.1
jsonschema==4.23.0
lxml==5.2.2
Markdown==3.6
mccabe==0.7.0
msrest==0.7.1
mypy-extensions==1.0.0
#mysqlclient==2.2.4
nodeenv==1.9.1
oauthlib==3.2.2
pathspec==0.12.1
Pillow==10.4.0
ply==3.11
pre-commit==3.7.1
psycopg2-binary==2.9.9
pyasn1==0.6.0
pycodestyle==2.12.0
pycparser==2.22
pycryptodomex==3.20.0
pyflakes==3.2.0
pyjwkest==1.4.2
PyJWT==2.8.0
pyproj==3.6.1
pyrsistent==0.20.0
python-dateutil==2.9.0.post0
python-jose==3.3.0
python3-openid==3.2.0
pytz==2024.1
PyYAML==6.0.2
regex==2024.5.15
requests==2.32.3
requests-oauthlib==2.0.0
rsa==4.9
six==1.16.0
social-auth-app-django==5.4.2
social-auth-core==4.5.4
soupsieve==2.5.0
sqlparse==0.5.1
toml==0.10.2
typed-ast==1.5.5
typing-extensions==4.12.2
tzlocal==5.2
urllib3==2.2.2
virtualenv==20.26.6
zipp==3.19.2
